Extract server endpoints for thumbnails and extensions into separate files
